+++ example/source/components/UI/Form.jsx
@@ -1,4 +1,4 @@
-import { createForm } from 're-validate';
+import { createForm } from 'react-form-validation';
 import validator from '../../validator';
 
 export default createForm(validator);