+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
 ## [Unreleased]
+- BREAKING: Rename package to "react-form-validation"
 - Add CHANGELOG